Create robots.demo.txt
This is for demo sites - we allow indexing since it's just demo data and we want the demo to be found!

You should probably NOT use this in corporate production environments. In production you should probably BLOCK search engines from indexing this information.  It could be a treasure trove of fishing and spammer data.  But the real solution is NOT to make this publicly accessible, since the spammers could scrape the information themselves even if it isn't index.  And there is no guarantee that search engines obey the the robots.txt.  This is just the LEAST I can do - YOU should do BETTER!